Time pass balaces

+++ src/bigcity/Person.java
@@ -4,7 +4,7 @@ public class Person {
 
     protected String name;
     protected int age;
-    protected int happiness;
+    protected double happiness;
     protected EducationLevel educationLevel;
     protected Zone home;
     protected Zone job;
@@ -21,7 +21,7 @@ public class Person {
         return age;
     }
 
-    public int getHappiness() {
+    public double getHappiness() {
         return happiness;
     }
 
@@ -33,7 +33,7 @@ public class Person {
         return male;
     }
 
-    public int changeHappinessBy(int value) {
+    public double changeHappinessBy(double value) {
         happiness += value;
         happinessGuard();
         return happiness;
